Restore XML serialization returning original xform submission
